In worst case scenario, (BIOS + RW main ISO + RW dongle), this setup is secure only if both the main ISO, the dongle and the machine cannot be tampered with; to reduce the attack surface on the dongle to the bootloader, GRUB currently first tries to boot the kernel from the encrypted dongle persistent partition and if not able to, from the ISO9660 partition.
This introduces a vulnerability in the sense that someone who recorded your signin having tampered with the bootloader and stolen/copied your dongle will be able to access the data on the dongle persistent partition.

There also is the issue that some machines require site-specific configurations to be able to have their keyboard working in the initrd.

Finally, there is currently no emergency mode, as both main iso + dongle are always needed to boot the system;

To solve all these issues, I am going to add two new partitions to the dongle ISO:
- a 'boot' partition encrypted with the same passphrase as the 'keys' image containing the 'airootfs' key unencrypted (to boot without initrd keyboard support and to keep the bottleneck at bootloader level);
- a 'recovery' partition encrypted with an entirely new passphrase to be kept as secure as the main ISO.

The above implies that:
- an evil maid who recorded an user sign-in and stole an user dongle will still need the main ISO to access the data;
- an evil maid who stole your main ISO won't have you lose your data.